Multiple Accounts

+++ db_interface/src/axolotl/PreKey.cpp
@@ -3,12 +3,13 @@
 using namespace std;
 using namespace sqlite;
 
-CriptextDB::PreKey CriptextDB::getPreKey(database db, short int id) {
+CriptextDB::PreKey CriptextDB::getPreKey(database db, int accountId, short int id) {
   string myPreKey;
   size_t myLen = 0;
-  db << "Select * from prekeyrecord where preKeyId == ?;"
+  db << "Select * from prekeyrecord where preKeyId == ? and accountId == ?;"
      << id
-     >> [&] (int preKeyId, string record, int recordLength) {
+     << accountId
+     >> [&] (int id, int preKeyId, string record, int recordLength) {
         myPreKey = record;
         myLen = (size_t)recordLength;
     };
@@ -24,12 +25,30 @@ CriptextDB::PreKey CriptextDB::getPreKey(database db, short int id) {
   return preKey;
 }
 
-bool CriptextDB::createPreKey(database db, short int id, char *keyRecord, size_t len) {
+bool CriptextDB::createPreKey(database db, int accountId, short int id, char *keyRecord, size_t len) {
   try {
-    db << "insert into prekeyrecord (preKeyId, record, recordLength) values (?,?,?);"
+    bool hasRow = false;
+    db << "begin;";
+    db << "Select preKeyId from prekeyrecord where preKeyId == ? and accountId == ?;"
      << id
-     << keyRecord
-     << static_cast<int>(len);
+     << accountId
+     >> [&] (string preKeyId) {
+        hasRow = true;
+    };
+    if (hasRow) {
+      db << "update prekeyrecord set record = ?, recordLength = ? where preKeyId == ? and accountId == ?;"
+        << keyRecord
+        << static_cast<int>(len)
+        << id
+        << accountId;
+    } else {
+      db << "insert into prekeyrecord (preKeyId, record, recordLength, accountId) values (?,?,?,?);"
+        << id
+        << keyRecord
+        << static_cast<int>(len)
+        << accountId;
+    }
+    db << "commit;";
     return true;
   } catch (exception& e) {
     std::cout << e.what() << std::endl;
@@ -37,10 +56,11 @@ bool CriptextDB::createPreKey(database db, short int id, char *keyRecord, size_t
   }
 }
 
-bool CriptextDB::deletePreKey(database db, short int id) {
+bool CriptextDB::deletePreKey(database db, int accountId, short int id) {
   try {
-    db << "delete from prekeyrecord where preKeyId == ?;"
-     << id;
+    db << "delete from prekeyrecord where preKeyId == ? and accountId == ?;"
+     << id
+     << accountId;
     return true;
   } catch (exception& e) {
     std::cout << e.what() << std::endl;
